requirements.txt
asgiref==3.4.1
autopep8==1.5.7
black==21.9b0
click==8.0.1
Django==3.2.5
djangorestframework==3.12.4
mypy-extensions==0.4.3
pathspec==0.9.0
pep8==1.7.1
platformdirs==2.4.0
psycopg2==2.9.1
pycodestyle==2.7.0
pymemcache==3.5.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.9.30
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
